+++ app/[teamId]/(index)/GroupTaskListWrapper.tsx
@@ -18,6 +18,7 @@ export type PointColorType =
   | 'yellow';
 
 interface GroupTaskListWrapperProps {
+  teamId: string | string[] | undefined;
   taskLists: ITaskList[] | null;
   onCreate: (params: _CreateTaskListParams) => void;
   onEdit: (params: _UpdateTaskListParams) => void;
@@ -35,6 +36,7 @@ const POINT_COLOR: PointColorType[] = [
 ];
 
 export default function GroupTaskListWrapper({
+  teamId,
   taskLists,
   onCreate,
   onEdit,
@@ -66,6 +68,7 @@ export default function GroupTaskListWrapper({
         taskLists.map((taskList, i) => (
           <GroupTaskList
             key={taskList.id}
+            teamId={teamId}
             taskList={taskList}
             pointColor={POINT_COLOR[i % POINT_COLOR.length]}
             onEdit={onEdit}
