+++ package.json
@@ -29,6 +29,7 @@
     "classnames": "^2.5.1",
     "clsx": "^2.1.1",
     "date-fns": "^3.0.0",
+    "dayjs": "^1.11.13",
     "lucide-react": "^0.471.2",
     "mdx-loader": "^3.0.2",
     "next": "15.1.4",
